+++ library/Swagger/Annotations/Resource.php
@@ -3,7 +3,7 @@ namespace Swagger\Annotations;
 
 /**
  * @license    http://www.apache.org/licenses/LICENSE-2.0
- *             Copyright [2013] [Robert Allen]
+ *             Copyright [2014] [Robert Allen]
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -125,7 +125,7 @@ class Resource extends AbstractAnnotation
             return false;
         }
         if (empty($this->resourcePath)) {
-            Logger::warning('@SWG\Resource() is missing "resourcePath" in '.AbstractAnnotation::$context);
+            Logger::warning('@SWG\Resource() is missing "resourcePath" in '.$this->_context);
             return false;
         }
         $this->apis = $apis;