+++ library/Swagger/Parser.php
@@ -3,7 +3,7 @@ namespace Swagger;
 
 /**
  * @license    http://www.apache.org/licenses/LICENSE-2.0
- *             Copyright [2013] [Robert Allen]
+ *             Copyright [2014] [Robert Allen]
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,10 +26,6 @@ use Doctrine\Common\Annotations\TokenParser;
 use Swagger\Annotations\AbstractAnnotation;
 use Swagger\Annotations\Model;
 use Swagger\Annotations\Resource;
-use Swagger\Contexts\ClassContext;
-use Swagger\Contexts\Context;
-use Swagger\Contexts\MethodContext;
-use Swagger\Contexts\PropertyContext;
 use Swagger\Processors\ProcessorInterface;
 
 /**
@@ -38,6 +34,11 @@ use Swagger\Processors\ProcessorInterface;
  */
 class Parser
 {
+    /**
+     * Allows Annotation classes to know the context of the annotation that is being processed.
+     * @var Context
+     */
+    public static $context;
 
     /**
      * The metadata about the API
@@ -57,12 +58,6 @@ class Parser
     protected $resources = array();
 
     /**
-     * Current resource
-     * @var Resource
-     */
-    protected $currentResource = false;
-
-    /**
      * All detected models
      * @var Model[]
      */
@@ -75,20 +70,9 @@ class Parser
     protected $partials = array();
 
     /**
-     * Current model
-     * @var Model
-     */
-    protected $currentModel = false;
-
-    /**
      * @var DocParser
      */
     private $docParser;
-
-    /**
-     * @var string
-     */
-    private $filename;
 
     /**
      * @var ProcessorInterface[]
@@ -116,7 +100,6 @@ class Parser
      */
     public function getResources()
     {
-        AbstractAnnotation::$context = $this->filename;
         $resources = array();
         foreach ($this->resources as $resource) {
             if ($resource->validate()) {
@@ -124,25 +107,7 @@ class Parser
             }
         }
         $this->resources = $resources;
-        AbstractAnnotation::$context = 'unknown';
         return $resources;
-    }
-
-    /**
-     * @param Resource $resource resource
-     */
-    public function appendResource(Resource $resource)
-    {
-        $this->resources[] = $resource;
-        $this->currentResource = $resource;
-    }
-
-    /**
-     * @return Annotation\Resource
-     */
-    public function getCurrentResource()
-    {
-        return $this->currentResource;
     }
 
     /**
@@ -151,7 +116,6 @@ class Parser
      */
     public function getModels()
     {
-        AbstractAnnotation::$context = $this->filename;
         $models = array();
         foreach ($this->models as $model) {
             if ($model->validate()) {
@@ -159,25 +123,7 @@ class Parser
             }
         }
         $this->models = $models;
-        AbstractAnnotation::$context = 'unknown';
         return $models;
-    }
-
-    /**
-     * @param Model $model model
-     */
-    public function appendModel(Model $model)
-    {
-        $this->models[] = $model;
-        $this->currentModel = $model;
-    }
-
-    /**
-     * @return Annotation\Model
-     */
-    public function getCurrentModel()
-    {
-        return $this->currentModel;
     }
 
     /**
@@ -217,17 +163,6 @@ class Parser
     }
 
     /**
-     * @param Annotations\Info $info
-     */
-    public function setInfo($info)
-    {
-        if ($this->info !== null) {
-            Logger::notice('Overwriting info with "'.$info->identity().'" found in '.AbstractAnnotation::$context);
-        }
-        $this->info = $info;
-    }
-
-    /**
      * @return Annotations\Authorizations
      */
     public function getAuthorizations() {
@@ -236,99 +171,97 @@ class Parser
     }
 
     /**
-     * @param Annotations\Authorizations $authorizations
-     */
-    public function setAuthorizations($authorizations) {
-        if ($this->authorizations !== null) {
-            Logger::notice('Overwriting authorizations with "'.$authorizations->identity().'" found in '.AbstractAnnotation::$context);
-        }
-        $this->authorizations = $authorizations;
-    }
-
-    /**
      * Extract and process all doc-comments from a file.
      * @param string $filename Path to a php file.
      */
     public function parseFile($filename)
     {
-        $this->filename = $filename;
-        $tokenParser = new TokenParser(file_get_contents($this->filename));
-        return $this->parseTokens($tokenParser);
+        $tokenParser = new TokenParser(file_get_contents($filename));
+        return $this->parseTokens($tokenParser, new Context(array('filename' => $filename)));
     }
 
     /**
      * Extract and process all doc-comments from the contents.
      * @param string $contents PHP code.
-     * @param string $context The original location of the contents.
+     * @param Context $context The original location of the contents.
      */
-    public function parseContents($contents, $context = 'unknown')
+    public function parseContents($contents, $context)
     {
-        $this->filename = $context;
         $tokenParser = new TokenParser($contents);
-        return $this->parseTokens($tokenParser);
+        return $this->parseTokens($tokenParser, $context);
     }
 
     /**
      * Shared implementation for parseFile() & parseContents().
      * @param TokenParser $tokenParser
      */
-    protected function parseTokens(TokenParser $tokenParser)
+    protected function parseTokens(TokenParser $tokenParser, $parseContext)
     {
         $this->docParser = new DocParser();
         $this->docParser->setIgnoreNotImportedAnnotations(true);
 
         $token = $tokenParser->next(false);
         $namespace = '';
-        $class = false;
-
         $imports = array(
             'swg' => 'Swagger\Annotations' // Use @SWG\* for swagger annotations (unless overwrittemn by a use statement)
         );
         $this->docParser->setImports($imports);
         $uses = array();
-        $docComment = false;
+        $classContext = $parseContext; // Use the parseContext until a classContext is created.
+        $comment = false;
+        $line = 0;
         while ($token != null) {
+            $previousToken = $token;
             $token = $tokenParser->next(false);
             if (is_array($token) === false) { // Ignore tokens like "{", "}", etc
                 continue;
             }
             if ($token[0] === T_DOC_COMMENT) {
-                $location = $this->filename . ' on line ' . $token[2];
-                AbstractAnnotation::$context = $location;
-                if ($docComment) { // 2 Doc-comments in succession?
-                    $this->parseContext(new Context($docComment));
+                if ($comment) { // 2 Doc-comments in succession?
+                    $this->parseContext(new Context(array('comment' => $comment, 'line' => $line), $classContext));
                 }
-                $docComment = $token[1];
+                $comment = $token[1];
+                $line = $token[2];
                 continue;
             }
             if ($token[0] === T_ABSTRACT) {
                 $token = $tokenParser->next(false); // Skip "abstract" keyword
             }
             if ($token[0] === T_CLASS) { // Doc-comment before a class?
+                if (is_array($previousToken) && $previousToken[0] === T_DOUBLE_COLON) {
+                    //php 5.5 class name resolution (i.e. ClassName::class)
+                    continue;
+                }
                 $token = $tokenParser->next();
-                $class = $namespace ? $namespace . '\\' . $token[1] : $token[1];
-                $this->currentModel = false;
+                $classContext = new Context(array(
+                    'class' => $namespace ? $namespace . '\\' . $token[1] : $token[1],
+                    'line' => $token[2],
+                ), $parseContext);
                 // @todo detect end-of-class and reset $class
-                if ($docComment) {
-                    $extends = null;
-                    $token = $tokenParser->next(false);
-                    if ($token[0] === T_EXTENDS) {
-                        $extends = $this->prefixNamespace($namespace, $tokenParser->parseClass(), $uses);
-                    }
-                    AbstractAnnotation::$context = $class . ' in ' . $location;
-                    $this->parseContext(new ClassContext($class, $extends, $docComment));
-                    $docComment = false;
+                $extends = null;
+                $token = $tokenParser->next(false);
+                if ($token[0] === T_EXTENDS) {
+                    $classContext->extends = $this->prefixNamespace($namespace, $tokenParser->parseClass(), $uses);
+                }
+                if ($comment) {
+                    $classContext->comment = $comment;
+                    $classContext->line = $line;
+                    $this->parseContext($classContext);
+                    $comment = false;
                     continue;
                 }
             }
-            if ($docComment) {
+            if ($comment) {
                 if ($token[0] == T_STATIC) {
                     $token = $tokenParser->next(false);
                     if ($token[0] === T_VARIABLE) { // static property
-                        AbstractAnnotation::$context = $class . '::' . $token[1] . ' in ' . $location;
-
-                        $this->parseContext(new PropertyContext(substr($token[1], 1), $docComment));
-                        $docComment = false;
+                        $this->parseContext(new Context(array(
+                            'property' => substr($token[1], 1),
+                            'static' => true,
+                            'comment' => $comment,
+                            'line' => $line
+                        ), $classContext));
+                        $comment = false;
                         continue;
                     }
                 }
@@ -338,30 +271,39 @@ class Parser
                         $token = $tokenParser->next(false);
                     }
                     if ($token[0] === T_VARIABLE) { // instance property
-                        AbstractAnnotation::$context = $class . '->' . substr($token[1], 1) . ' in ' . $location;
-                        $this->parseContext(new PropertyContext(substr($token[1], 1), $docComment));
-                        $docComment = false;
+                        $this->parseContext(new Context(array(
+                            'property' => substr($token[1], 1),
+                            'comment' => $comment,
+                            'line' => $line
+                        ), $classContext));
+                        $comment = false;
                     } elseif ($token[0] === T_FUNCTION) {
                         $token = $tokenParser->next(false);
                         if ($token[0] === T_STRING) {
-                            AbstractAnnotation::$context = $class . '->' . $token[1] . '(...)' . ' in ' . $location;
-                            $this->parseContext(new MethodContext($token[1], $docComment));
-                            $docComment = false;
+                            $this->parseContext(new Context(array(
+                                'method' => $token[1],
+                                'comment' => $comment,
+                                'line' => $line
+                            ), $classContext));
+                            $comment = false;
                         }
                     }
                     continue;
                 } elseif ($token[0] === T_FUNCTION) {
                     $token = $tokenParser->next(false);
                     if ($token[0] === T_STRING) {
-                        AbstractAnnotation::$context = $class . '->' . $token[1] . '(...)' . ' in ' . $location;
-                        $this->parseContext(new MethodContext($token[1], $docComment));
-                        $docComment = false;
+                        $this->parseContext(new Context(array(
+                            'method' => $token[1],
+                            'comment' => $comment,
+                            'line' => $line
+                        ), $classContext));
+                        $comment = false;
                     }
                 }
                 if (in_array($token[0], array(T_NAMESPACE, T_USE)) === false) { // Skip "use" & "namespace" to prevent "never imported" warnings)
                     // Not a doc-comment for a class, property or method?
-                    $this->parseContext(new Context($docComment));
-                    $docComment = false;
+                    $this->parseContext(new Context(array('comment' => $comment, 'line' => $line), $classContext));
+                    $comment = false;
                 }
             }
             if ($token[0] === T_NAMESPACE) {
@@ -381,31 +323,58 @@ class Parser
                 continue;
             }
         }
-        if ($docComment) { // File ends with a T_DOC_COMMENT
-            $this->parseContext(new Context($docComment));
+        if ($comment) { // File ends with a T_DOC_COMMENT
+            $this->parseContext(new Context(array('comment' => $comment, 'line' => $line), $classContext));
         }
-        AbstractAnnotation::$context = 'unknown';
+        $rootContext = $parseContext->getRootContext();
+        if ($rootContext->info) {
+            $this->info = $rootContext->info;
+        }
+        if ($rootContext->authorizations) {
+            $this->authorizations = $rootContext->authorizations;
+        }
     }
 
     /**
      *
-     * @param Context $context Content containing the docComment
+     * @param Context $context
      * @return AbstractAnnotation[]
      */
     protected function parseContext($context)
     {
         try {
-            $annotations = $this->docParser->parse($context->getDocComment(), AbstractAnnotation::$context);
+            self::$context = $context;
+            $annotations = $this->docParser->parse($context->comment, $context);
+            self::$context = null;
         } catch (\Exception $e) {
-            Logger::warning($e);
+            self::$context = null;
+            if (preg_match('/^(.+) at position ([0-9]+) in '.preg_quote($context, '/').'\.$/', $e->getMessage(), $matches)) {
+                $errorMessage = $matches[1];
+                $errorPos = $matches[2];
+                $atPos = strpos($context->comment, '@');
+                $context->line += substr_count($context->comment, "\n", 0, $atPos + $errorPos);
+                $lines = explode("\n", substr($context->comment, $atPos, $errorPos));
+                $context->character = strlen(array_pop($lines)) + 1; // position starts at 0 character starts at 1
+                Logger::warning(new \Exception($errorMessage.' in '.$context, $e->getCode(), $e));
+            } else {
+                Logger::warning($e);
+            }
             return array();
         }
 
         foreach ($annotations as $annotation) {
             foreach ($this->processors as $processor) {
-                if ($processor->supports($annotation, $context)) {
-                    $processor->process($this, $annotation, $context);
+                $processor->process($annotation, $context);
+            }
+            if ($annotation->hasPartialId()) {
+                if ($this->hasPartial($annotation->_partialId)) {
+                    Logger::notice('partial="' . $annotation->_partialId . '" is not unique. another was found in ' . $annotation->_context);
                 }
+                $this->setPartial($annotation->_partialId, $annotation);
+            } elseif ($annotation instanceof Resource) {
+                $this->resources[] = $annotation;
+            } elseif ($annotation instanceof Model) {
+                $this->models[] = $annotation;
             }
         }
         return $annotations;