+++ CHANGELOG.md
@@ -6,6 +6,22 @@ Changelog
 
 * Implemented @SWG\Info & @SWG\Authorizations which can be used to augment the api-docs.json with metadata.
 
+## 0.9.4
+
+* Fixed regressions #145 & #146
+
+## 0.9.3
+
+* New Context/ProcessorInterface allows access to the parse context via the Annotation #142
+* Added context to DocParser error messages and position is converted back into line and character information. #124
+* Fixed model inheritance bug #140.
+* getResourceList() now has a template option, which accepts a json file which can contain the `info` and `authentication` fields. #139
+* The `--api-doc-template` option will use the imported `swaggerVersion`, `apiVersion` and `basePath` as defaults.
+
+## 0.9.2
+
+* Fixed issue with PHP 5.5  class name resolution (ClassName::class)
+
 ## 0.9.1
 
 * Generate ".{format}" instead of ".json" suffix when using the commandline.